Final example, formatting and import cleanup

+++ streamlit_app.py
@@ -1,16 +1,19 @@
-from socket import create_connection
-import streamlit as st
-import duckdb
-from urllib.request import urlretrieve
-import pyarrow.dataset as ds
+import os
+from datetime import datetime
 from time import time
+from urllib.request import urlretrieve
+
+import altair as alt
+import duckdb
+import streamlit as st
 
 
 @st.cache
 def download_dataset():
     url = "https://github.com/cwida/duckdb-data/releases/download/v1.0/lineitemsf1.snappy.parquet"
     dst = "lineitemsf1.snappy.parquet"
-    urlretrieve(url, dst)
+    if not os.path.exists(dst):
+        urlretrieve(url, dst)
     return dst
 
 
@@ -29,7 +32,7 @@ st.title("DuckDB Demo")
 
 st.write(
     """
-This is a simple demo showing off a few features of DuckDB (duckdb.org).
+This is a simple demo showing off a few features of DuckDB (https://duckdb.org).
 
 If the application is booting cold, it needs a few more seconds to download the dataset. 
 After that, all queries are run directly on the parquet file downloaded.
@@ -72,3 +75,33 @@ run_timed_query(
         AND l_quantity < 24; """,
     connection,
 )
+
+st.header("Shipped totals")
+st.write("A small example showing interactive filtering")
+
+start_date, end_date = st.slider(
+    "Date range for analysis",
+    value=(datetime(1992, 1, 1, 9, 30), datetime(1999, 1, 1, 9, 30)),
+    format="DD/MM/YY",
+)
+
+results = connection.execute(
+    """SELECT l_shipdate, sum(l_extendedprice * l_discount) AS revenue
+    FROM
+        'lineitemsf1.snappy.parquet'
+    WHERE
+        l_shipdate >= CAST(? AS date)
+        AND l_shipdate < CAST(? AS date)
+    group by l_shipdate
+    order by l_shipdate asc
+    """,
+    [start_date, end_date],
+).fetchdf()
+
+c = (
+    alt.Chart(results)
+    .mark_line()
+    .encode(x=alt.X("l_shipdate:T", timeUnit="yearmonthdate"), y="revenue")
+)
+
+st.altair_chart(c, use_container_width=True)
